refactor(users): relocate repo interfaces to application/usecases (DIP × 3)
v0.160.1 polish Item 1 — closes 1 of 7 Tier 2 carry-forwards from
v0.160.0 (PR #288, closes #283).

UserProfileRepository, DepartmentRepository, PositionRepository
moved from `internal/modules/users/domain/repositories/` к
`internal/modules/users/application/usecases/repository_interfaces.go`
(consolidated) per Clean Architecture DIP gate ("Repository interfaces
— в пакете-потребителе (`usecase/`), НЕ в `domain/`"). Mirror к
v0.157.1 curriculum + v0.162.1 messaging precedent.

UserAccountRepository уже жил в this package since v0.139.0 — the
relocation completes the parity. UserFilter query DTO остаётся в
domain/repositories/user_profile_repository.go (now contains only
that value type — sentinels + query DTOs pattern, mirror curriculum
precedent where filters stay в domain).

Zero behavior change:
- 3 PG impls (department_repository.go, position_repository.go,
  user_profile_repository.go) switch named import к usersUsecases.
- 3 usecase structs drop the `repositories.` prefix on field +
  constructor param types (structural typing — concrete impls still
  satisfy).
- main.go switches `usersRepositories.UserProfileRepository`
  references к `usersUsecases.UserProfileRepository`; the alias stays
  для UserFilter struct usage on line 292.

All users tests + main.go cmd build green post-relocation.

Refs #313
